clean up code smells, fix tests
